Updates LogPlugin to use Protected for $logAdapter
Changes $logAdapter in the LogPlugin from private to protected. This
makes it easier for people to extend from LogPlugin to make any needed
customizations and still have access to $logAdapter. Without this
change, if someone wanted to use $logAdapter, they will need to
duplicate the entire class instead of extending.